refactor: page, layout 코드 수정
